fix application properties parse

+++ app/src/main/java/direct/thither/onamove/pages/SearchSettings.java
@@ -145,7 +145,7 @@ public class SearchSettings {
             @Override
             public void onProgressChanged(SeekBar elem, int v, boolean state){
                 String v1 = Integer.toString(v);
-                m_props.update_freq = Long.parseLong(v1);
+                m_props.update_freq = (long)Double.parseDouble(v1);
                 m_props.apply_pref("update_freq", m_props.update_freq);
                 update_freq_holder.setText(v1+"s");
                 m_props.renders.get("1").set_timers();
